// WADNR.Web/src/app/shared/generated/model/relationship-type-grid-row.ts
export class RelationshipTypeGridRow { 
    RelationshipTypeID?: number;
    RelationshipTypeName?: string | null;
    CanStewardProjects?: boolean;
    IsPrimaryContact?: boolean;
    CanOnlyBeRelatedOnceToAProject?: boolean;
    ShowOnFactSheet?: boolean;
    ReportInAccomplishmentsDashboard?: boolean;
    RelationshipTypeDescription?: string | null;
    AssociatedOrganizationTypeNames?: Array<string> | null;
    ProjectOrganizationCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}