// WADNR.Web/src/app/shared/generated/model/relationship-type-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class RelationshipTypeUpsertRequest { 
    RelationshipTypeName?: string | null;
    RelationshipTypeDescription?: string | null;
    CanStewardProjects?: boolean;
    IsPrimaryContact?: boolean;
    CanOnlyBeRelatedOnceToAProject?: boolean;
    ShowOnFactSheet?: boolean;
    ReportInAccomplishmentsDashboard?: boolean;
    OrganizationTypeIDs?: Array<number> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface RelationshipTypeUpsertRequestForm { 
    RelationshipTypeName?: FormControl<string>;
    RelationshipTypeDescription?: FormControl<string>;
    CanStewardProjects?: FormControl<boolean>;
    IsPrimaryContact?: FormControl<boolean>;
    CanOnlyBeRelatedOnceToAProject?: FormControl<boolean>;
    ShowOnFactSheet?: FormControl<boolean>;
    ReportInAccomplishmentsDashboard?: FormControl<boolean>;
    OrganizationTypeIDs?: FormControl<Array<number>>;
}

export class RelationshipTypeUpsertRequestFormControls { 
    public static RelationshipTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RelationshipTypeDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CanStewardProjects = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsPrimaryContact = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CanOnlyBeRelatedOnceToAProject = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ShowOnFactSheet = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ReportInAccomplishmentsDashboard = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationTypeIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}